Upgrade libuv to f5ff8694

+++ deps/uv/test/test-hrtime.c
@@ -46,7 +46,7 @@ TEST_IMPL(hrtime) {
 
   diff = b - a;
 
-  printf("diff = %llu\n", diff);
+  printf("diff = %llu\n", (unsigned long long int)diff);
 
   ASSERT(diff >= NANOSEC / MICROSEC);
   ASSERT(diff > MICROSEC);
